requirements.txt
asgiref==3.2.10
Django==3.0.7
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
pylint==2.5.3
djangorestframework==3.11.1
markdown==3.2.2
django-filter==2.3.0
